simplify AccessFor signature
